Problem Statement: When gprecoverseg is running gpstate -e command displays status of the recovered segments. if the user kills the gprecoverseg process then gpstate -e continues to report the status of the old gprecoverseg.
Root Cause
gpstate -e command checks if the recovery_progress_file exists or not before displaying the gprecoverseg progress.
Since the recovery_progress_file is not removed when the gprecoverseg is forcefully killed, gpstate -e will continue to display the progress.

Fix:
To fix this issue, gpstate -e code is modified to verify if ~/workspace/gpdb/gpAux/gpdemo/datadirs/qddir/demoDataDir-1/gprecoverseg.lock exists before reporting the gprecoverseg progress. Since the user is aware that the lock file should be removed if gprecoverseg process is killed, the double check ensures that gpstate -e does not report the progress of the old killed gprecoverseg.