When gprecoverseg is running gpstate -e command displays status of the recovered segments. if the user kills the gprecoverseg process then gpstate -e continues to report the status of the old gprecoverseg. As a code fix, previosuly we added a check to see if ~/workspace/gpdb/gpAux/gpdemo/datadirs/qddir/demoDataDir-1/gprecoverseg.lock exists before reporting the gprecoverseg progress. This fix has a limitation as in the user needs to manually delete the lock file, otherwise gpstate -e will continue to report the progress of gprecoverseg. To avoid this, now we are also checking if gprecoverseg process is running and gpstate -e reports the progress only if this check succeeds.
Some of the behave test cases had to be removed/modified to accomodate this code change.
Previosuly some of the gpstate -e test cases used to create a sample recovery_progress_file and check the format of the gpstate -e output. This test case worked before as the gprecoverseg process was not running.
Now since there is a gprecoverseg process running which updates the recovery_progress file, this test cases will always fail.
Hence 1 test case is removed and 2 are modified.